+++ core/src/main/scala/ai/lum/odinson/lucene/index/OdinsonIndex.scala
@@ -12,8 +12,23 @@ import com.typesafe.config.Config
 import org.apache.lucene.analysis.Analyzer
 import org.apache.lucene.analysis.core.WhitespaceAnalyzer
 import org.apache.lucene.document.{ Document => LuceneDocument }
-import org.apache.lucene.index.{ Fields, IndexReader }
-import org.apache.lucene.search.{ Collector, CollectorManager, Query, TopDocs }
+import org.apache.lucene.index.{ Fields, IndexReader, Term }
+import org.apache.lucene.search.join.{
+  QueryBitSetProducer,
+  ScoreMode,
+  ToChildBlockJoinQuery,
+  ToParentBlockJoinQuery
+}
+import org.apache.lucene.search.{
+  BooleanClause,
+  BooleanQuery,
+  Collector,
+  CollectorManager,
+  MatchAllDocsQuery,
+  Query,
+  TermQuery,
+  TopDocs
+}
 import org.apache.lucene.store.{ Directory, FSDirectory, IOContext, RAMDirectory }
 
 import java.nio.file.Paths
@@ -42,8 +57,57 @@ trait OdinsonIndex {
 
   def indexOdinsonDoc(doc: OdinsonDocument): Unit
 
+  /** Removes all `org.apache.lucene.document.Document`s representing an [[ai.lum.odinson.Document]] (including metadata).
+    * @param odinsonDocId The ID of the Odinson Document to remove from the index.
+    */
+  def deleteOdinsonDoc(odinsonDocId: String): Unit
+
+  /** Updates index entries for an [[ai.lum.odinson.Document]] by first removing all `org.apache.lucene.document.Document`s representing an [[ai.lum.odinson.Document]] (including metadata) with the same ID.
+    * @param odinsonDocId The ID of the Odinson Document to update from the index.
+    */
+  def updateOdinsonDoc(doc: OdinsonDocument): Unit
+
+  /** Creates an `org.apache.lucene.search.Query` matching the final document in a block representing some [[ai.lum.odinson.Document]].  The match is based on the `ai.lum.odinson.Document.id` field.
+    * @param odinsonDocId The ID of the Odinson Document associated with a block of Lucene documents.
+    */
+  def mkDocIdQueryFor(odinsonDocId: String): Query = {
+    // query to identify the last lucene doc in the block
+    val queryBuilder = new BooleanQuery.Builder()
+    queryBuilder.add(
+      new BooleanClause(
+        new TermQuery(new Term(OdinsonIndexWriter.DOC_ID_FIELD, odinsonDocId)),
+        BooleanClause.Occur.MUST
+      )
+    )
+    queryBuilder.build()
+  }
+
+  /** Creates an `org.apache.lucene.search.Query` matching all but the final document in a block representing some [[ai.lum.odinson.Document]] (i.e., the children of that block).  The match is based on the `ai.lum.odinson.Document.id` field.
+    * @param odinsonDocId The ID of the Odinson Document associated with a block of Lucene documents.
+    */
+  def mkChildrenQueryFor(odinsonDocId: String): Query = {
+    val parentQuery = mkDocIdQueryFor(odinsonDocId)
+    val parentsFilter = new QueryBitSetProducer(parentQuery)
+    // identify all children in a doc based on a field in the final doc
+    new ToChildBlockJoinQuery(parentQuery, parentsFilter)
+  }
+
+  /** Collects IDs for all `org.apache.lucene.document.Document`s representing some [[ai.lum.odinson.Document]] in the index.
+    * @param odinsonDocId The ID of the Odinson Document for which to collect all of its component `org.apache.lucene.document.Document` IDs.
+    */
+  def luceneDocIdsFor(odinsonDocId: String): Seq[Int] = {
+    // final lucene doc in block
+    val parentQuery = mkDocIdQueryFor(odinsonDocId)
+    // all but final lucene doc in block
+    val childrenQuery = mkChildrenQueryFor(odinsonDocId)
+    val res = search(childrenQuery).scoreDocs ++ search(parentQuery).scoreDocs
+    // get IDs
+    res.map(_.doc)
+  }
+
   def write(block: java.util.Collection[LuceneDocument]): Unit
 
+  // FIXME: use a constant value for default value of `limit` that represents largest possible value
   def search(query: Query, limit: Int = 1000000000): TopDocs
 
   def search[CollectorType <: Collector, ResultType](
@@ -64,7 +128,7 @@ trait OdinsonIndex {
 
   def doc(docId: Int): LuceneDocument
 
-  def doc(docID: Int, fieldNames: Set[String]): LuceneDocument
+  def doc(docId: Int, fieldNames: Set[String]): LuceneDocument
 
   def lazyIdGetter(luceneDocId: Int): LazyIdGetter
 
@@ -130,7 +194,7 @@ object OdinsonIndex {
     val displayField = config.apply[String]("odinson.displayField")
     // Always store the display field, also store these additional fields
     if (!storedFields.contains(displayField)) {
-      throw new OdinsonException("`odinson.index.storedFields` must contain `odinson.displayField`")
+      throw OdinsonException("`odinson.index.storedFields` must contain `odinson.displayField`")
     }
 
     val computeTotalHits = config.apply[Boolean]("odinson.computeTotalHits")
@@ -166,5 +230,9 @@ object OdinsonIndex {
     )
 
   }
+
+  /** Context manager to ensure index gets closed.
+    */
+  def usingIndex[T](config: Config)(f: OdinsonIndex => T): T = using(fromConfig(config))(f)
 
 }