+++ core/src/test/scala/ai/lum/odinson/lucene/index/TestIncrementalIndex.scala
@@ -1,8 +1,10 @@
 package ai.lum.odinson.lucene.index
 
+import ai.lum.odinson.metadata.MetadataCompiler
 import ai.lum.odinson.test.utils.OdinsonTest
 import ai.lum.odinson.utils.IndexSettings
 import ai.lum.odinson.utils.exceptions.OdinsonException
+import ai.lum.odinson.{ ExtractorEngine, TokensField }
 import com.typesafe.config.{ Config, ConfigValueFactory }
 import org.apache.lucene.store.FSDirectory
 import org.scalatest.BeforeAndAfterEach
@@ -91,35 +93,160 @@ class TestIncrementalIndex extends OdinsonTest with BeforeAndAfterEach {
   }
 
   it should "incrementally add Odinson Documents to an open index" in {
-    val index = OdinsonIndex.fromConfig(testConfig)
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      val aliens = getDocument("alien-species")
+      index.indexOdinsonDoc(aliens)
 
-    val aliens = getDocument("alien-species")
-    index.indexOdinsonDoc(aliens)
+      index.numDocs() shouldBe 2 // # of lucenen docs
 
-    index.numDocs() shouldBe 2 // # of lucenen docs
+      val gummyBears = getDocument("gummy-bears-consumption")
+      index.indexOdinsonDoc(gummyBears)
 
-    val gummyBears = getDocument("gummy-bears-consumption")
-    index.indexOdinsonDoc(gummyBears)
-
-    index.numDocs() shouldBe 4 // # number of lucene docs
-    index.close()
+      index.numDocs() shouldBe 4 // # number of lucene docs
+    }
   }
 
   it should "incrementally write Odinson Documents to a previously closed index" in {
-    var index = OdinsonIndex.fromConfig(testConfig)
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      val aliens = getDocument("alien-species")
+      index.indexOdinsonDoc(aliens)
 
-    val aliens = getDocument("alien-species")
-    index.indexOdinsonDoc(aliens)
+      index.numDocs() shouldBe 2 // # of lucene docs
+    }
 
-    index.numDocs() shouldBe 2 // # of lucene docs
-    index.close()
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      val gummyBears = getDocument("gummy-bears-consumption")
+      index.indexOdinsonDoc(gummyBears)
 
-    index = OdinsonIndex.fromConfig(testConfig)
-    val gummyBears = getDocument("gummy-bears-consumption")
-    index.indexOdinsonDoc(gummyBears)
-
-    index.numDocs() shouldBe 4 // # number of lucene docs
-    index.close()
+      index.numDocs() shouldBe 4 // # number of lucene docs
+    }
   }
 
+  it should "incrementally delete Odinson Documents from an open index" in {
+    // doc w/ 1 sentence & metadata w/ 2 sets of nested fields
+    val pies = getDocument("tp-pies")
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      index.indexOdinsonDoc(pies)
+
+      val odinsonDocId = pies.id
+      index.numDocs() shouldBe 4 // # of lucene docs
+      index.luceneDocIdsFor(odinsonDocId).size shouldBe 4
+
+      // Deleting an Odinson document should delete all of its sentences, as well as its metadata (including any nested fields)
+      index.deleteOdinsonDoc(odinsonDocId)
+
+      index.numDocs() shouldBe 0
+      index.luceneDocIdsFor(odinsonDocId).size shouldBe 0
+    }
+  }
+
+  it should "incrementally delete Odinson Documents from a previously closed index" in {
+    // doc w/ 1 sentence & metadata w/ 2 sets of nested fields
+    val pies = getDocument("tp-pies")
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      index.indexOdinsonDoc(pies)
+
+      index.numDocs() shouldBe 4 // # of lucene docs
+    }
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      index.numDocs() shouldBe 4 // # of lucene docs
+      index.luceneDocIdsFor(pies.id).size shouldBe 4
+
+      // Deleting an Odinson document should delete all of its sentences, as well as its metadata (including any nested fields)
+      index.deleteOdinsonDoc(pies.id)
+
+      index.numDocs() shouldBe 0
+      index.luceneDocIdsFor(pies.id).size shouldBe 0
+    }
+  }
+
+  it should "incrementally delete an Odinson Document from a previously closed index containing multiple documents" in {
+    // doc w/ 1 sentence & metadata w/ 2 sets of nested fields
+    // we'll index and delete this one
+    val pies = getDocument("tp-pies")
+    // we'll index this one, but not delete it
+    val briggs = getDocument("tp-briggs")
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      index.indexOdinsonDoc(pies)
+      index.numDocs() shouldBe 4 // # of lucene docs
+      index.indexOdinsonDoc(briggs)
+      index.numDocs() should be > 4
+    }
+
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      val totalDocs = index.numDocs()
+      totalDocs should be > 4
+      index.luceneDocIdsFor(pies.id).size shouldBe 4
+      // Deleting an Odinson document should delete all of its sentences, as well as its metadata (including any nested fields)
+      index.deleteOdinsonDoc(pies.id)
+
+      val remaining = totalDocs - 4
+      index.numDocs() shouldBe remaining
+      index.luceneDocIdsFor(pies.id).size shouldBe 0
+    }
+  }
+
+  it should "not crash if asked to incrementally delete a non-existent Odinson Document" in {
+    OdinsonIndex.usingIndex(testConfig) { index =>
+      // the index is empty
+      index.numDocs() shouldBe 0
+      // while no such doc exists,
+      // this should not cause an error
+      noException should be thrownBy index.deleteOdinsonDoc("tp-pies")
+    }
+  }
+
+  it should "update an index with a new version of an Odinson Document (respecting changes to sentences)" in {
+    // index doc and then update with an abbreviated version
+    ExtractorEngine.usingEngine(testConfig) { engine =>
+      // we'll index this doc and later update
+      val majorBriggs = getDocument("tp-briggs")
+      val index = engine.index
+      // the index is empty
+      index.numDocs() shouldBe 0
+      index.indexOdinsonDoc(majorBriggs)
+      val oldCount = index.numDocs()
+      // let's remove some sentences and update the index
+      val minorBriggs = majorBriggs.copy(sentences = Seq(majorBriggs.sentences.head))
+      index.updateOdinsonDoc(minorBriggs)
+      index.numDocs() should be < oldCount
+    }
+  }
+
+  it should "update an index with a new version of an Odinson Document (respecting changes to metadata)" in {
+    // index doc and then update with additional metadata
+    ExtractorEngine.usingEngine(testConfig) { engine =>
+      // "This must be where pies go to die."
+      // we'll index this doc and later update
+      val pies = getDocument("tp-pies")
+      val metadataFilter = MetadataCompiler.mkQuery("flavor contains 'cherry'")
+      val query = engine.mkFilteredQuery("[lemma=pie]", metadataFilter)
+      val index = engine.index
+      // the index is empty
+      index.numDocs() shouldBe 0
+      index.indexOdinsonDoc(pies)
+      // the query shouldn't match
+      engine.query(query).scoreDocs.length shouldBe 0
+
+      // let's replace that doc with one that will match the query
+      val flavorField = TokensField(name = "flavor", tokens = Seq("cherry"))
+      val cherryPies = pies.copy(metadata = pies.metadata ++ Seq(flavorField))
+      index.updateOdinsonDoc(cherryPies)
+      // the query should now match
+      val res = engine.query(query)
+      res.scoreDocs.length shouldBe 1
+    }
+  }
+
+  it should "not crash if asked to update a non-existent Odinson Document" in {
+    ExtractorEngine.usingEngine(testConfig) { engine =>
+      val pies = getDocument("tp-pies")
+      val index = engine.index
+      // the index is empty
+      index.numDocs() shouldBe 0
+      // while no such doc exists,
+      // this should not cause an error
+      noException should be thrownBy index.updateOdinsonDoc(pies)
+    }
+  }
 }